+++ groups/bsl/bslma/bslma_autodeallocator.t.cpp
@@ -1468,7 +1468,8 @@ int main(int argc, char *argv[])
 
         for (int i = 0; i < NUM_TEST; ++i) {
             memory[i] = ta.allocate(sizeof(int) * (i + 1));
-            ASSERT(sizeof(int) * (i + 1) * (i + 2) / 2 == TA.numBytesInUse());
+            ASSERT(sizeof(int) * (i + 1) * (i + 2) / 2 ==
+                   (size_t) TA.numBytesInUse());
         }
 
         ASSERT(sizeof(int) * NUM_TEST * (NUM_TEST + 1) / 2
@@ -1487,7 +1488,8 @@ int main(int argc, char *argv[])
         char *cmemory[NUM_TEST];
         for (int i = 0; i < NUM_TEST; ++i) {
             cmemory[i] = (char *)ta.allocate(sizeof(int) * (i + 1));
-            ASSERT(sizeof(int) * (i + 1) * (i + 2) / 2 == TA.numBytesInUse());
+            ASSERT(sizeof(int) * (i + 1) * (i + 2) / 2 ==
+                   (size_t) TA.numBytesInUse());
         }
 
         ASSERT(sizeof(int) * NUM_TEST * (NUM_TEST + 1) / 2
